+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Added
 
 - Add a link to README to error message which appears in case you do not specify neither `before` nor `after` parameters.
+- Add a warning message if `allowRootAndOutside` option is on and if there is unsafe removing.
 
 
 # 1.0.0 (June 5, 2018)